device_manager/requirements.txt
django-environ==0.11.1
channels==3.0.4
channels-redis==3.3.1
msgpack-python==0.5.6
paho-mqtt==1.6.1
six==1.16.0
sqlparse==0.4.4
uvicorn[standard]
gunicorn==19.1.1
mysqlclient